Address review comments: Split functionality to get, transform, load format Added data samples
