+++ cartography/sync.py
@@ -8,7 +8,7 @@ import cartography.intel.analysis
 import cartography.intel.aws
 import cartography.intel.gcp
 import cartography.intel.create_indexes
-import cartography.intel.crxcavator
+import cartography.intel.crxcavator.crxcavator
 
 
 logger = logging.getLogger(__name__)
@@ -141,9 +141,9 @@ def build_default_sync():
     sync = Sync()
     sync.add_stages([
         ('create-indexes', cartography.intel.create_indexes.run),
-        ('aws', cartography.intel.aws.start_aws_ingestion),
-        ('gcp', cartography.intel.gcp.start_gcp_ingestion),
+        # ('aws', cartography.intel.aws.start_aws_ingestion),
+        # ('gcp', cartography.intel.gcp.start_gcp_ingestion),
         ('crxcavator', cartography.intel.crxcavator.start_extension_ingestion),
-        ('analysis', cartography.intel.analysis.run),
+        # ('analysis', cartography.intel.analysis.run),
     ])
     return sync
