+++ tests/data/crxcavator/crxcavator.py
@@ -1,0 +1,51 @@
+# flake8: noqa
+REPORT_RESPONSE = {
+    'ppedokobpbdajgiejhnjfbdjlgobcpkp',
+    [
+        {'data':
+            {
+                'risk':
+                    {
+                        'total': 437,
+                        'metadata': {}
+                    },
+                'webstore':
+                    {
+                        'address': '',
+                        'email': '',
+                        'icon': 'https://lh3.googleusercontent.com/6vvUTv5SIaggk-mliDV6o_CfcDIIKu7MR5gL8Ep5QtODWrYOPj_hgywZQ-FImelit_Id1thKNSU=w128-h128-e365',
+                        'last_updated': '2016-02-22', 'name': 'f*ck overlays',
+                        'offered_by': '',
+                        'permission_warnings': [
+                            'Your data on all websites'
+                        ],
+                        'privacy_policy': '',
+                        'rating': 4.6778846,
+                        'rating_users': 208,
+                        'short_description': 'adds a context menu to delete sign-up/sign-in overlays that prevent you from viewing a page',
+                        'size': '13.95KiB',
+                        'support_site': '',
+                        'users': 38241,
+                        'version': '',
+                        'website': '',
+                        'type': 'Extension',
+                        'price': ''
+                    }
+            },
+            'extension_id': 'ppedokobpbdajgiejhnjfbdjlgobcpkp',
+            'version': '1.1.2'
+        }
+    ]
+}
+
+USER_RESPONSE = {
+    'ppmacakdbohpklnhpaecpoohbkapggpo',
+    {
+        '1.7':
+            {
+                'name': 'MindMeister Quick Access',
+                'users':
+                    ['user@example.com']
+            }
+    }
+}
